Move base rollup config into new config location
