+++ .npmignore
@@ -7,7 +7,7 @@ ambient.d.ts
 CHANGELOG.md
 design.md
 Gruntfile.js
-rollup.config.js
 tsconfig.json
 tslint.json
 typings.json
+config
