+++ packages/apollo-cache/rollup.config.js
@@ -1,6 +1,6 @@
 import resolve from 'rollup-plugin-node-resolve';
 
-import build from '../../rollup.config';
+import build from '../../config/rollup.config';
 
 export default build('apollo.cache.core', {
   plugins: [resolve()],
