+++ packages/apollo-client/rollup.config.js
@@ -1,3 +1,3 @@
-import build from '../../rollup.config';
+import build from '../../config/rollup.config';
 
 export default build('apollo.core');
